requirements.txt
fastapi
uvicorn
pydantic
openai
faiss-cpu
sqlite3
python-dotenv
requests
numpy
pandas 